requirements.txt
cachetools==4.0.0
certifi==2019.11.28
chardet==3.0.4
Click==7.0
dialogflow==0.7.2
Flask==1.1.1
google-api-core==1.16.0
google-auth==1.11.2
googleapis-common-protos==1.51.0
grpcio==1.27.2
gunicorn==20.0.4
idna==2.9
itsdangerous==1.1.0
Jinja2==2.11.1
MarkupSafe==1.1.1
protobuf==3.11.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyJWT==1.7.1
PySocks==1.7.1
pytz==2019.3
requests==2.23.0
rsa==4.0
six==1.14.0
twilio==6.35.5
urllib3==1.25.8
utils==1.0.1
Werkzeug==1.0.0
